Changed for pip

+++ TD_BRAIN_code/BRAIN_code/alphaPowerandiAPF.py
@@ -6,8 +6,9 @@ Created on Mon Oct 26 11:55:28 2020
 @author: hannekevandijk
 """
 #%% 1
-from BCD_gitprojects.preprocessing_beta.io import FilepathFinder
-from BCD_gitprojects.preprocessing_beta.interprocessing import interdataset as intds
+#from BCD_gitprojects.preprocessing_beta.io import FilepathFinder
+from BRAIN_code.inout import FilepathFinder
+from BRAIN_code.interprocessing import interdataset as intds
 import numpy as np
 import pickle5 as pickle
 import os
@@ -16,6 +17,7 @@ import numpy as np
 
 root_dir = '/NAS/database/BCD_OA_database/'
 exclude = ['BAD', '._' ,'Apple']
+
 files = FilepathFinder('.npy', root_dir+'BCD_OA_preprocessed_recoded_trial_set/', exclude =exclude)
 files.__get_filenames__()
 
@@ -25,7 +27,7 @@ def computeFFT(data):
     from scipy.signal import hann
     hannwin = np.array(hann(data.shape[-1]))
     power = np.squeeze(np.mean(np.abs(np.fft.fft(data*hannwin))**2,axis=0))
-    freqs = np.linspace(0,varargs['fs']/2,np.int(len(power)/2)) #power1
+    freqs = np.linspace(0,varargs['fs']/2,np.int_(len(power)/2)) #power1
     return power[:len(freqs)], freqs
 
 
@@ -55,7 +57,7 @@ for f in files.selectedfiles:
     output['session'].append(sessid)
     output['conds'].append(f.rsplit('_')[-2])
     #read belonging tsvdat data
-    idx = np.where((tsvdat['subID']==np.int(sid)) & (tsvdat['sessID']==np.int(sessid)))[0][0]
+    idx = np.where((tsvdat['subID']==np.int_(sid)) & (tsvdat['sessID']==np.int_(sessid)))[0][0]
     output['age'].append(tsvdat['age'][idx])
     output['sex'].append(tsvdat['gender'][idx])
     output['power'] = np.vstack((output['power'],np.squeeze(computeFFT(chandata)[0])))
